ejecutable/index/_internal/requirements.txt
asgiref==3.7.2
CTkMessagebox==2.5
customtkinter==5.2.2
darkdetect==0.8.0
et-xmlfile==1.1.0
greenlet==3.0.3
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
Pillow==10.0.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
tkcalendar==1.6.1
typing_extensions==4.11.0
tzdata==2023.3
